[MPP-25615] Replace libgpla with custom functions.
This removes the dependency to libblas, liblapack and libf2c. We only used
three functions from those libraries:

dgemv: used to perform Matrix * Vector multiplication.
ddot: used to calculate dot product of two vectors
dgesdd: used to calculate Singular Value Decomposition of a matrix.

The first two were trivial to replace; the replacements were about 20 lines
in total. dgesdd was much more complicated. I picked an SVD implementation
from JAMA, a linear algebra library for Java, and adapted it to C. The
JAMA code uses the same SVD algorithm as LAPACK, so this should produce same
results (except for rounding). JAMA is in public domain.

Now that we're in control of the calculation for the whole time, also put
a few CHECK_FOR_INTERRUPTS() calls in the function. That makes it possible
to interrupt the pinv() function, which could run for minutes with large
inputs.